Version should be a string rather than json.Number
For certain version 'strings' it might work out to use the type
json.Number, but this isn't always the case. The best example is a
version with a patch level, like 1.12.1 (as nginx might report it). This
value isn't allowed to be serialized back into JSON as it isn't a
number, it's a string.

Long story short, the Version member should be a string rather than a
number.